swapped hljs with netitor in ref code-samples
