+++ www/data/references/css-reference-cascade.html
@@ -5,7 +5,7 @@
 
 <h2>what takes precedence?</h2>
 <p>You can copy and paste the code below into the editor to experiment with it and see what the rendered output looks like.</p>
-<code-example name="css-cascade-ex-1"></code-example>
+<code-sample name="css-cascade-ex-1"></code-sample>
 <ul>
   <li>
     <b>The Order</b>:<br>
@@ -24,11 +24,11 @@
 </ul>
 
 <h2>other considerations</h2>
-<code-example name="css-cascade-ex-2"></code-example>
+<code-sample name="css-cascade-ex-2"></code-sample>
 <p>
   In the example above the text will be purple, regardless of what was defined in the <i>.red-text { }</i> CSS rule because the inline styling always takes precedence over rules in an attached stylesheet or between <i>&lt;style&gt;</i>tags.
 </p>
-<code-example name="css-cascade-ex-3"></code-example>
+<code-sample name="css-cascade-ex-3"></code-sample>
 <p>
   Notice that there's a space between "red" and "green" in the example above, this means there are two different classes in the stylesheet <i>.red { }</i> and <i>.green { }</i>. Both list of properties will be combined and applied to the <i>div</i>, but if there is a conflict, then the general "Last One" rule applies, whichever one of these two classes is defined last in the stylesheet takes precedence.
 </p>
