+++ www/data/references/css-reference-main.html
@@ -288,14 +288,14 @@
 <h2>CSS Selectors</h2>
 <p>
   The example above uses a type selector, which means it will apply the rules to every <code>&lt;a&gt;</code> tag on the page. You can apply the same rule to multiple elements by including them in a comma separated list within the selector. For example:
-  <code-example name="css-ex-1"></code-example>
+  <code-sample name="css-ex-1"></code-sample>
   This will change the color of all the <code>&lt;h1&gt;</code>, <code>&lt;h2&gt;</code>, and <code>&lt;h3&gt;</code> tags on the page to purple.
 </p>
 <br>
 <p>
   There are many other kinds of selectors which give you different ways of targeting specific elements to apply your CSS rules to. The most common selector is likely to be the class selector which targets any element whose class attribute’s value matches that of the class name.
 </p>
-<code-example name="css-ex-2"></code-example>
+<code-sample name="css-ex-2"></code-sample>
 <br>
 <p>
   There are a number of different kinds of selectors and thus a number of ways to specify which elements on a page you want to apply a CSS rule to), The most common are:
