+++ www/index.html
@@ -17,6 +17,9 @@
    ` _ _ _ _ _ __ _ _ _ _ _ _ _ _ _ _  ,/
                                     .'`
                               ( ◕ ◞ ◕ )つ
+
+  to learn more about how this site was built check out our GitHub wiki:
+  https://github.com/netizenorg/netnet.studio/wiki/Project-Architecture
     -->
 
     <!-- social media stuff -->
@@ -40,7 +43,6 @@
 
     <link rel="stylesheet" href="css/text-bubbles.css"> <!-- <text-bubble> -->
     <link rel="stylesheet" href="css/reference-widget.css"> <!-- <widget-slide> -->
-    <link rel="stylesheet" href="css/libs/hljs-default.css"> <!-- <code-example> -->
   </head>
   <body>
 
@@ -61,7 +63,6 @@
     <script src="./FileUploader.js"></script>
     <script src="./netitor.min.js"></script>
     <script src="./js/libs/fuse.js"></script>
-    <script src="./js/libs/highlight.pack.js"></script>
 
     <script src="./js/Convo.js"></script>
     <script src="./js/Widget.js"></script>
